update code: content provider to room

+++ app/src/main/java/annin/my/android/popularmovies2/data/MovieDao.java
@@ -19,7 +19,7 @@ import annin.my.android.popularmovies2.pojo.Movie;
         LiveData<List<Movie>> loadAllMovies();
 
         @Insert
-        int insertMovie(Movie movieEntry);
+        long insertMovie(Movie movieEntry);
 
         @Update(onConflict = OnConflictStrategy.REPLACE)
         void updateMovie(Movie movieEntry);
